refactoring presentor screeen, body and details
