+++ lib/features/presentor/ui/presentor_screen.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'package:desktop_window/desktop_window.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:styled_widget/styled_widget.dart';
 
@@ -12,7 +13,9 @@ import '../../../common/utils/constants/app_assets.dart';
 import '../../../common/widgets/action/fab_widget.dart';
 import '../../../common/widgets/presentor/presentor_animate.dart';
 import '../../../common/widgets/progress/custom_snackbar.dart';
+import '../../../common/widgets/progress/general_progress.dart';
 import '../../../core/theme/theme_colors.dart';
+import '../bloc/presentor_bloc.dart';
 import '../common/presentor_intents.dart';
 import '../common/presentor_utils.dart';
 import '../common/slide_utils.dart';
@@ -20,7 +23,7 @@ import '../common/slide_utils.dart';
 part 'widgets/presentor_view.dart';
 part 'widgets/fab_widget.dart';
 part 'widgets/presentor_body.dart';
-part 'widgets/presentor_pageview.dart';
+part 'widgets/presentor_details.dart';
 part 'widgets/presentor_tabbar.dart';
 
 class PresentorScreen extends StatefulWidget {
@@ -41,6 +44,7 @@ class PresentorScreen extends StatefulWidget {
 
 class PresentorScreenState extends State<PresentorScreen> {
   late AppLocalizations l10n;
+  late SongExt song;
   late String songTitle, songBook;
   bool hasChorus = false, likeChanged = false, slideHorizontal = false;
   List<String> songVerses = [];
@@ -51,14 +55,14 @@ class PresentorScreenState extends State<PresentorScreen> {
   @override
   void initState() {
     super.initState();
-    if (widget.song.content.contains("CHORUS")) {
+    song = widget.song;
+    if (song.content.contains("CHORUS")) {
       hasChorus = true;
     }
-    l10n = AppLocalizations.of(context)!;
-    songTitle = songItemTitle(widget.song.songNo, widget.song.title);
-    songBook = refineTitle(widget.song.songbook);
+    songTitle = songItemTitle(song.songNo, widget.song.title);
+    songBook = refineTitle(song.songbook);
     if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
-      DesktopWindow.setFullScreen(true);
+      //DesktopWindow.setFullScreen(true);
     }
   }
 
@@ -94,43 +98,43 @@ class PresentorScreenState extends State<PresentorScreen> {
     }
   }
 
-  void _toggleLike() {
-    setState(() {
-      widget.song.liked = !widget.song.liked;
-      likeChanged = true;
-    });
-
-    String message = widget.song.liked ? l10n.songLiked : l10n.songDisliked;
-
-    CustomSnackbar.show(context, message, isSuccess: widget.song.liked);
-  }
-
   @override
   Widget build(BuildContext context) {
-    return PopScope(
-      canPop: false,
-      onPopInvokedWithResult: (bool didPop, dynamic result) async {
-        if (didPop) return;
-        if (context.mounted) {
-          Navigator.pop(context, likeChanged);
-        }
-      },
-      child: Scaffold(
-        appBar: AppBar(
-          title: Text(songTitle),
-          actions: [
-            Tooltip(
-              message: widget.song.liked ? l10n.songLike : l10n.songDislike,
-              child: IconButton(
-                onPressed: _toggleLike,
-                icon: Icon(
-                    widget.song.liked ? Icons.favorite : Icons.favorite_border),
-              ),
-            ),
-          ],
-        ),
-        body: PresentorBody(parent: this),
-        floatingActionButton: PresentorFabWidget(song: widget.song),
+    l10n = AppLocalizations.of(context)!;
+
+    return BlocProvider(
+      create: (context) => PresentorBloc()..add(LoadSong(widget.song)),
+      child: BlocConsumer<PresentorBloc, PresentorState>(
+        listener: (context, state) {
+          if (state is PresentorFailureState) {
+            CustomSnackbar.show(
+              context,
+              feedbackMessage(state.feedback, l10n),
+            );
+          } else if (state is PresentorLikedState) {
+            setState(() {
+              song.liked = !song.liked;
+              likeChanged = true;
+            });
+            if (state.liked) {
+              CustomSnackbar.show(context, l10n.songLiked, isSuccess: true);
+            } else {
+              CustomSnackbar.show(context, l10n.songDisliked);
+            }
+          } else if (state is PresentorLoadedState) {
+            setState(() {
+              songVerses = state.songVerses!;
+              widgetTabs = state.widgetTabs!;
+              widgetContent = state.widgetContent!;
+            });
+          }
+        },
+        builder: (context, state) {
+          return state.maybeWhen(
+            progress: () => Scaffold(body: CircularProgress()),
+            orElse: () => PresentorBody(parent: this),
+          );
+        },
       ),
     );
   }
